// projects/ngx-crontab-builder/src/lib/models/crontab-mode.ts
export enum CrontabMode {
  EVERY_X_SECONDS = "EVERY_X_SECONDS",
  EVERY_X_MINUTES = "EVERY_X_MINUTES",
  EVERY_X_HOURS = "EVERY_X_HOURS",
  EVERY_DAY = "EVERY_DAY",
  EVERY_WEEK = "EVERY_WEEK",
  NTH_WEEKDAY = "NTH_WEEKDAY",
  EVERY_MONTH = "EVERY_MONTH",
  LAST_DAY_OF_MONTH = "LAST_DAY_OF_MONTH",
  CUSTOM = "CUSTOM"
}